requirements.txt
tensorflow==1.8.0
numpy==1.14.1
selenium==3.11.0
tushare==1.2.4
requests==2.18.4
jieba==0.39
pymongo==3.6.1
Pillow==5.2.0
beautifulsoup4==4.6.0
scikit_learn==0.19.1
